requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.8.0
docstring-parser==0.15
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.2
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
instructor==0.4.8
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.3
litellm==1.20.9
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.4
numpy==1.26.1
ollama==0.1.5
openai==1.1.1
opencv-python==4.8.1.78
orjson==3.9.12
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pydantic==2.4.2
pydantic_core==2.10.1
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rq==1.15.1
rubicon-objc==0.4.7
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tiktoken==0.5.2
tokenizers==0.15.1
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.22.0
wcwidth==0.2.10
websockets==12.0
yarl==1.9.4
zipp==3.17.0
